Move group.[ch] to mutt/
