+++ Makefile.autosetup
@@ -64,7 +64,7 @@ NEOMUTT=	neomutt$(EXEEXT)
 NEOMUTTOBJS=	account.o addrbook.o alias.o bcache.o browser.o color.o commands.o \
 		complete.o compose.o compress.o conststrings.o copy.o \
 		curs_lib.o edit.o editmsg.o enriched.o enter.o \
-		filter.o flags.o git_ver.o group.o handler.o hdrline.o help.o hook.o \
+		filter.o flags.o git_ver.o handler.o hdrline.o help.o hook.o \
 		index.o init.o keymap.o mailbox.o main.o menu.o muttlib.o \
 		mutt_account.o mutt_attach.o mutt_body.o mutt_header.o \
 		mutt_history.o mutt_logging.o mutt_parse.o mutt_signal.o \
@@ -249,7 +249,7 @@ ALLOBJS+=	$(LIBEMAILOBJS)
 # libmutt
 LIBMUTT=	libmutt.a
 LIBMUTTOBJS=	mutt/base64.o mutt/buffer.o mutt/charset.o mutt/date.o \
-		mutt/envlist.o mutt/exit.o mutt/file.o mutt/hash.o \
+		mutt/envlist.o mutt/exit.o mutt/file.o mutt/group.o mutt/hash.o \
 		mutt/history.o mutt/list.o mutt/logging.o mutt/mapping.o \
 		mutt/mbyte.o mutt/md5.o mutt/memory.o mutt/path.o mutt/regex.o \
 		mutt/sha1.o mutt/signal.o mutt/string.o
