+++ mutt/group.c
@@ -33,8 +33,13 @@
 #include "mutt/mutt.h"
 #include "email/lib.h"
 #include "group.h"
-#include "globals.h"
-#include "send.h"
+#include "hash.h"
+#include "logging.h"
+#include "memory.h"
+#include "regex3.h"
+#include "string2.h"
+
+struct Hash *Groups;
 
 /**
  * mutt_pattern_group - Match a pattern to a Group
@@ -160,7 +165,7 @@ static void group_add_addrlist(struct Group *g, struct Address *a)
     ;
 
   q = mutt_addr_copy_list(a, false);
-  q = mutt_remove_xrefs(g->as, q);
+  q = mutt_addr_remove_xrefs(g->as, q);
   *p = q;
 }
