+++ mutt/hash.h
@@ -26,7 +26,7 @@
 
 #include <stdbool.h>
 #include <stdint.h>
-#include <stdio.h>
+#include <stdlib.h>
 
 /**
  * union HashKey - The data item stored in a HashElem
